tests/test_engagement_tracking.py
```python
import unittest
from src import engagement_tracking

class TestEngagementTracking(unittest.TestCase):

    def setUp(self):
        self.engagement_tracking = engagement_tracking.EngagementTracking()
        self.user_data = {
            "id": 1,
            "name": "John Doe",
            "email": "johndoe@example.com"
        }
        self.event_data = {
            "id": 1,
            "name": "Virtual Concert",
            "date": "2022-12-31"
        }

    def test_track_engagement(self):
        result = self.engagement_tracking.track_engagement(self.user_data, self.event_data)
        self.assertTrue(result)

    def test_get_engagement_data(self):
        self.engagement_tracking.track_engagement(self.user_data, self.event_data)
        engagement_data = self.engagement_tracking.get_engagement_data(self.user_data["id"], self.event_data["id"])
        self.assertIsNotNone(engagement_data)

    def test_get_event_analytics(self):
        self.engagement_tracking.track_engagement(self.user_data, self.event_data)
        event_analytics = self.engagement_tracking.get_event_analytics(self.event_data["id"])
        self.assertIsNotNone(event_analytics)

if __name__ == '__main__':
    unittest.main()
```